Retrieve JSON data from an external source
Allow the user to define a custom URL to retrieve the JSON
data (instead of hitting Read the Docs API) needed to add the banner
into the page.

This is useful when there is a git tag that doesn't want to be
updated, but now the warning message is out of date. So, by defining a
custom URL (potentially a JSON under the git repository served by
github, for example) this file is retrieved by the documentation and
displayed the custom message needed at certain date.

The JSON file can be updated independently from the documentation/tag/etc.